get component
