+++ src/test/scala/nl.knaw.dans.easy.bagstore/BagStoreGetSpec.scala
@@ -21,23 +21,24 @@ import org.apache.commons.io.FileUtils
 
 import scala.util.Success
 
-class BagStoreGetSpec extends BagStoreFixture with BagStoreGet with BagStoreAddComponent with BagStoreOutputContext {
+class BagStoreGetSpec extends BagStoreFixture with BagStoreGetComponent with BagStoreAddComponent with BagStoreOutputContext {
   val outputBagPermissions: String = "rwxrwxrwx"
   private val TEST_BAGS_DIR = Paths.get("src/test/resources/bags")
   private val TEST_BAGS_PRUNED = TEST_BAGS_DIR.resolve("basic-sequence-pruned")
   private val TEST_BAG_PRUNED_A = TEST_BAGS_PRUNED.resolve("a")
 
   override val add = new BagStoreAdd {}
+  override val get = new BagStoreGet {}
 
   "get" should "return exactly the same Bag as was added" in {
-    val result = get(add.add(TEST_BAG_PRUNED_A).get, testDir)
+    val result = get.get(add.add(TEST_BAG_PRUNED_A).get, testDir)
     result shouldBe a[Success[_]]
     pathsEqual(TEST_BAG_PRUNED_A, testDir.resolve("a")) shouldBe true
   }
 
   it should "create Bag base directory with the name of parameter 'output' if 'output' does not point to existing directory" in {
     val output = testDir.resolve("non-existent-directory-that-will-become-base-dir-of-exported-Bag")
-    val result = get(add.add(TEST_BAG_PRUNED_A).get, output)
+    val result = get.get(add.add(TEST_BAG_PRUNED_A).get, output)
     result shouldBe a[Success[_]]
     pathsEqual(TEST_BAG_PRUNED_A, output) shouldBe true
   }
@@ -45,7 +46,7 @@ class BagStoreGetSpec extends BagStoreFixture with BagStoreGet with BagStoreAddC
   it should "return a File in the Bag that was added" in {
     val bagId = add.add(TEST_BAG_PRUNED_A).get
     val fileId = FileId(bagId, Paths.get("data/x"))
-    val result = get(fileId, testDir)
+    val result = get.get(fileId, testDir)
     result shouldBe a[Success[_]]
     pathsEqual(TEST_BAG_PRUNED_A.resolve("data/x"), testDir.resolve("x")) shouldBe true
   }
@@ -53,7 +54,7 @@ class BagStoreGetSpec extends BagStoreFixture with BagStoreGet with BagStoreAddC
   it should "rename a File to name specified in 'output' if 'output' does not point to an existing directory" in {
     val bagId = add.add(TEST_BAG_PRUNED_A).get
     val fileId = FileId(bagId, Paths.get("data/x"))
-    val result = get(fileId, testDir.resolve("x-renamed"))
+    val result = get.get(fileId, testDir.resolve("x-renamed"))
     result shouldBe a[Success[_]]
     // Attention: pathsEqual cannot be used, as it also compares file names
     FileUtils.contentEquals(TEST_BAG_PRUNED_A.resolve("data/x").toFile, testDir.resolve("x-renamed").toFile) shouldBe true
